+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   `get_cache_path()`: Gets the path to a cache folder.
+-   `ValueExtractor()` - A tool for extracting values from a set of possible entries.
 
 ## [0.8.1] - 2023-10-03